CDMS-1181 - Updates API client and refines JSON deserialization (#303)
* Updates API client and refines JSON deserialization

Upgrades the `TradeImportsDataApi.Api.Client` package to a newer version. This update introduces changes that require a refinement in how incoming JSON messages are deserialized.

Explicitly sets the JSON `PropertyNamingPolicy` to `CamelCase` to ensure robust and consistent handling of message properties, especially regarding their casing. This helps maintain compatibility and prevents deserialization issues with the updated client's data structures.

Includes a minor adjustment to a log message to reflect the correct casing of a property from the updated client.

Relates to CDMS-1181

* updated de serialization

* Update Defra.TradeImportsDataApi.Api.Client version

---------

Co-authored-by: Chris Johnson <chrisjohnson@equalexperts.com>